+++ lib/expandable_search_bar_plus.dart
@@ -1,1 +1,2 @@
 export 'src/expandable_search_bar_plus.dart';
+export 'src/expandable_search_bar_plus_controller.dart';